+++ lib/pact_api/pact_api_v1.dart
@@ -58,7 +58,7 @@ class PactApiV1 extends IPactApiV1 {
               'signatureValidation': signatureValidation,
             },
           );
-    print(uri.toString());
+    // print(uri.toString());
     // print('local');
     // print(jsonEncode(command));
     http.Response response = await http.post(